initial changes to project
